pyproject.toml
[tool.poetry]
name = "aviauth-api"
version = "0.1.0"
description = "Authentication microservice based on FastAPI"
authors = ["Alex Rozum <rozumalex@gmail.com>"]
license = "GNU General Public License v3.0"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.70.1"
uvicorn = "^0.16.0"
toml = "^0.10.2"
SQLAlchemy = "^1.4.29"
psycopg2 = "^2.9.2"
alembic = "^1.7.5"
sqlmodel = "^0.0.5"
asyncpg = "^0.25.0"
greenlet = "^1.1.2"
tenacity = "^8.0.1"
structlog = "^21.5.0"
fastapi-versioning = "^0.10.0"
python-multipart = "^0.0.5"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^3.2.0"
email-validator = "^1.1.3"
sendgrid = "^6.9.3"
httpx = "^0.21.1"
pytest-asyncio = "^0.16.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.16.0"
commitizen = "^2.20.3"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
pytest-mock = "^3.6.1"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
tag_format = "$version"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
